Adopt React-Redux to App
