+++ src/components/Cell/Cell.tsx
@@ -2,7 +2,7 @@ import * as $ from 'jquery';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import { ActionDispatcher } from '../Table/Container';
+import { ActionDispatcher } from '../Store/Store';
 
 interface CellProps {
   key: string;
